Make &version a generator.
The first result is the string from version.h (exactly as before).
Subsequent results give details from the build (the output from
git describe --long HEAD) and the name of the branch that the software
was built on. The Unicon compiler uses the information to build a
more informative output when the -version option is given. If no tags
are defined, the git hash of the HEAD revision is used instead of the
output of git describe (which fails if there are no tags).