Allow more flexibility when tunrning features on/off
The configure script now differentiates between implicit and explicit options
when configuring features. Explicit options will cause the script to stop
with an error message if one of the dependencies is missing. Example:

./configure --enable-voip

stops and reports the following error if libvoip is missing

configure: error: "voip requires libvoip-dev"

Whereas:

./configure

Will skip the voip feature if libvoip-dev is not present.

Signed-off-by: Jafar Al-Gharaibeh <to.jafar@gmail.com>