Always use the unvarnished branch name when building.
If the branch name is the same name as a tag (which will happen on
long term maintenance branches) using git rev-parse to extract the
branch name results in extra information to make it "non-ambiguous".
Using git branch --show-current always returns the branch name with
nothing added: Unfortunately, "old gits" don't support it, so sed
is used to remove the unwanted information.