Update langprocs.icn procedure object_to_string
A change is required in object_to_string to change the field index
from a value of 3 to 1 to ensure that the correct mapping is obtained.

Indexing the object fields to avoid __s and __m is not needed. The first
field to be seen as obj[1] is the first field defined in the class
declaration.

Signed-off-by: Bruce Rennie <brennie@dcsi.net.au>