Minor updates to union.icn and scan_util.icn
In scan_util.icn, the code for moveMatch() is cleaned up and the header comment adjusted to reflect the actual behavior.
In union.icn the decoding of class and record UniON strings is simplified and a potential conflict arrising from conflicts between user-defined class fields and the internal class fields is addressed.
Signed-off-by: Steve Wampler <sbw@tapestry.tucson.az.us>