Suppress clang warnings produced by rtt code for is:thing
clang complains about the extra parentheses output (by rtt) when emitting code
for is:null etc.. The best thing would be to modify rtt so that the extra
parentheses were only emitted if they were really necessary. The easiest thing
would be to shut clang up about it globally. The easy thing is to shut clang up
only during the is:thing code. We do the easy thing.