Add a bit manipulation plugin.
The routines are the equivalent of the Unicon built-in functions
except they never produce a large integer and there is no special
case for the sign bit.
   bor      bitwise inclusive or       (ior)
   bxor     bitwise exclusive or       (ixor)
   band     bitwise and                (iand)
   bcom     bitwise one's complement   (icom)
   bshift   bitwise shift              (ishift)
plus
   bits     enquiry and bit extraction
   bit      single bit extraction
   brot     bit rotation

bit(x,n) works the same way as string indexing (but indexing bits instead
of characters, including negative indices meaning "from the end").
  bit(x,1) is the ls bit of x,      (analogous to s[1]).
  bit(x,n) is the nth bit of x      (analogous to s[n] when n is +ve).
  bit(x,0) is the ms bit of x       (analogous to s[0]).
  bit(x,-n) is analogous to s[-n] i.e indexes from the ms end of x.

bits(x,n,m) works the same way as substring indexing s[n:m].
Note that, as with substrings, the bits are not reversed if m < n.
   bits(x,n,m) = bits(x,m,n)

The routines are not portable by design (i.e. on machines with different
word lengths they may give different results), which is the entire point.