Add control over the size and contents of the heap verifier log buffer.
The log buffer size may be varied by defining VRFY_LOGSIZE in define.h
The "x is verified" log messages are controlled by an environment
variable VRFYOP which has a bit set for each type that the "is verified"
message is desired. (e.g. for records the value is 1<<T_Record = 128).