Fix key and data display of testgdbm and testndbm.
Some (perhaps most?) key and data pairs in uniclass.dir are not null
terminated (there is a dsize field, which is correct, but testgdbm and
testndbm didn't use it) resulting in extra garbage characters being
displayed after the key and data. The fix is to use a precision spec
in the printf format strings.